allow for user to resend a webhook if being rate limited
allows a user to pass an optional boolean kwarg of rate_limit_retry. If the value
of the boolean is True, then the program will automatically attempt to
resend the webhook as per the value returned from the retry limit error
message of the Discord Webhook response until the webhook is sent
successfully.

Also refactored the sending of the execution of the post request to make
the code a little bit nicer for the purposes of this new feature.